Fix a caption update issue when a status message is being displayed.
The caption string is not updated immediately if a status message is
being displayed. When the status message is erased after the timeout,
not all of the caption string is redrawn. As a result, if the caption
had changed while the status message was being displayed, it's not
updated properly. Reported in #23945. This change fixes the bug.