Nested list of windows in the windowlist view
The -g flag to windowlist will cause the windowlist to be nested, i.e.
the windows belonging to a group window in the list will show up
indented on the left. The -g flag is compatible with the -m flag.

Also, in a windowlist, pressing  or  will jump to its parent group
window.

References #23856