Require password, when applicable, for remote detach
If a session has a password set, then it should be necessary for the
user to enter the password before a remote detach. This change achieves
that behaviour.

A small change in the messaging system is necessary, thus bumping the
message version, for the first time since it has been introduced, it
seems. But I don't think it's going to break anything.

Discussion at #24029