Use the last created window as foreground window
Some commands don't work correctly if screen is started in daemon mode
and not attached/detached at least once. This happens because without
any attaches, no window is considered as the foreground window when
remotely executing a command without a display. In such a case, treat
the last created window as the foreground window, as suggested in
 #10074.