Prompt for input for 'stuff' command if necessary
"
This feature is useful when using a very slow remote shell session (SSH
over a bad GPRS connection). By locally preparing a command you can
enter a command and send it as bulk (without typos ;-)).
"

The gain from this patch is that, instead of doing ^a:stuff "command",
the user would be doing ^aXcommand. This is a relatively pretty minor
convenience, but I suppose it could be useful if the user needs to do
this frequently.