Allow highlighting bell/monitor windows in caption
The command is 'rendition'. Details in man-page. The defaults are
currently set in a way to make sure that the new settings are
noticeable. This changeset is preferred over either of the patches in
savannah #18382, because it allows specifying the renditions for both
bell and monitor windows, and is flexible enough that new renditions
can be added if desired.