'screen -Q <command>' to get back results.
This is the just the start of querying an existing session. The goal
is to allow something like "windows=`screen -Q windows`" in a bash
script to get back the results from the 'windows' command.

Most of the framework is done. Now the commands need to be updated
to specially deal with the queries.