Set both Xterm title and icon name.
screen's so-called "rxvt_osc" option (actually xterm control sequences)
includes handling for OSC parameters 0 and 1.  Referring to ctlseqs.ms,
these parameters are of interest:
    Ps = 0  -> Change Icon Name and Window Title to Pt
    Ps = 1  -> Change Icon Name to Pt
    Ps = 2  -> Change Window Title to Pt
screen allows 0 and 1 (it should allow 2, but it is not used as often,
and screen's code is unpalatable).  However, it maps the 0 into 1,
setting only the icon name.  That isn't useful and is opposite to the
normal sense these control sequences would be used in.

Closes Debian#344759

http://www.xfree86.org/current/ctlseqs.html