Refresh cleverly to improve scrolling speed.
This should fix a lot of complaints about slow scrolling in vertical
splits. The idea is to update the display once at the end of a screen
write, rather than a lot of time during the write, which was the root
of the slowness.

It is possible that LayPauseUpdateRegion needs to be called in a few
more places, but this works for now for the tests I have done.