Allow adding visual notification in caption string.
'%P' in caption string will set '%?' to true if the current region is
in copy/paste mode. So, if you want to simply change the color etc. of
your caption, prepend '%?%P%{XXX}%?' to your caption and you're set ('XXX'
stands for the usual attribute/color modifier). If, on the other hand,
you want to change the caption string in copy mode, change your caption
string to '%?%PCaption for copy mode%:Your usual caption%?'.