Fix cursor positioning after a search in copy mode
The problem would show up with splits (both vertical and
horizontal), because the layer wouldn't remember the position the
cursor was moved to, and when screen refreshes the regions (if
there's more than one), it takes the cursor to the old position. The
fix is to simply remember the cursor position in the layer.