Complete the query-response system.
Some commands now can be queried from a remote session using the
'-Q' flag, e.g. 'screen -Q windows'. The commands will send the
response to the stdout of the querying process. If there was an
error in the command, then the querying process will exit with
a non-zero status.

The commands that can be queried now are:
  echo
  info
  lastmsg
  time
  title
  windows