Reset displays before dumping a core.
If defined SHADOWPW, we may have passwd records in core, that the user
would not be able to access otherwise. In that case, we should not dump
core, as the core file would contain the passwd records, and would be
readable for the user.
We do not explicitly check for eff_uid == 0, because if his real_uid is
also 0 he could have read all this anyway. Leaving only the cases where the
two uids differ.