Use intelligent screen refresh only in split regions.
When there are no splits, pausing output on a layer isn't necessary,
since it doesn't hit the performance. But we do want to continue to
pause output when there are splits to improve scrolling performance.
Fixes savannah bug #29055. Thanks to Kazuo Teramoto for reporting
the bug and testing the fix.