Fix build with -Werror=format-security
Some distributions ( mandriva for example ) use
-Werror=format-security by default when build C software to try to
enhance code quality and prevent security bugs. As the policy is to
correct all occurences of the error ( so there is no false positive,
nor question about to fix it or not ), a patch that fix the
error on gnu screen 4.0.3 have been added to the package. I have
rediffed it against latest git snapshot, to be sent upstream.

It mainly add "%s" where applicable.