Fix refresh when double-cell characters end a line
Double cell characters that end a line are not properly displayed
when there are split regions. Takeshi Banse detected the problem and
a fix for it in Savannag bug #29106. A better fix for this was
applied in the unicode++ branch, but that won't work in 'master'.
This is approximately the same fix, in a slightly different way.