Fix 'hardcopy -h'.
Make -h save scrollback with or without a file name given. Also,
preserve the unlikely case in which someone actually does want to
save display to '-h'; with

  :hardcopy -- -h

or, to save both history and display to '-h'

  :hardcopy -h -h

Also,

  :hardcopy --
  :hardcopy -h --

will, as they always have, save display or history and display to
'--'.