Use NEED_FORE flag for 'hardcopy'.
Screen wouldn't do anything with 'hardcopy' if there was no
foreground window, except overwriting the specified file (when
hardcopy_append is off, of course). But the NEED_FORE flag wasn't
set. I am not sure why this was the case. So I have added the
NEED_FORE flag so that 'hardcopy' command fails if there is no
foreground window.