Improve cursor store/restore on smcup/rmcup.
When going to alternate screen buffer on smcup, the terminal is
supposed to store the cursor position, rendition etc. Screen used to
store only the position (and that was just a little buggy anyway).
So change this behaviour to store all information about the cursor,
and restore everything on rmcup.

Apparently, the terminal is supposed to store the cursor for smcup
(CSI ? 1049 h) and DECSC (ESC 7, CSI s, CSI ? 1048 h) separately. So
be it.