Fix segfault when restoring w_alt.cursor.
This fixes a crash when the window is made too small (zero height).
The crash can be reproduced by turning on altscreen and caption, and
resizing the terminal window to height = 1. After this, if the
termianl window height is increased, screen would crash when going
out of the alt screen.