Ensure that the title prompt only appears once
Windows in screen can be renamed using C-a A (title). This can be done
repeatedly so that any number of prompts for a new window name overlay
each other. Once the user entered a name in the topmost prompt and
presses enter, he has to do it again for all the prompts underneath.
This can be very confusing in case the user hit C-a A several times
without noticing or realizing.

Signed-off-by: Amadeusz Sławiński <amade@asmblr.net>